# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNILIV Admin API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: getRooms
      tags: [rooms]
      summary: List rooms
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/PropertyIdParam"
      responses:
        "200":
          description: Rooms list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomsListResponse"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update room
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomResponse"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete room
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── RESIDENTS ────────────────────────────────────────
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List users
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - in: query
          name: role
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ANNOUNCEMENTS ────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
    PropertyIdParam:
      name: propertyId
      in: query
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ─── AUTH ─────────────────────────────────────────────
...
    UserDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
        propertyId:
          type: string
          nullable: true
        isActive:
          type: boolean
        lastLogin:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, isActive, createdAt]
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/UserDto"
      required: [success, data]
    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
        propertyId:
          type: string
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        propertyId:
          type: string
        isActive:
          type: boolean

    # ─── DASHBOARD ────────────────────────────────────────
...
    RoomDto:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        number:
          type: string
        floor:
          type: integer
        wing:
          type: string
          nullable: true
        type:
          type: string
        capacity:
          type: integer
        status:
          type: string
        occupancy:
          type: integer
        createdAt:
          type: string
      required: [id, propertyId, number, floor, type, capacity, status, occupancy, createdAt]
    RoomResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/RoomDto"
      required: [success, data]
    RoomsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/RoomDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreateRoomBody:
      type: object
      properties:
        propertyId:
          type: string
        number:
          type: string
        floor:
          type: integer
        wing:
          type: string
        type:
          type: string
        capacity:
          type: integer
        status:
          type: string
      required: [propertyId, number, floor, type, capacity]

    # ─── RESIDENT ─────────────────────────────────────────
...