# lib/api-zod/src/generated/types/roomsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNILIV Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaginationMeta } from "./paginationMeta";
import type { RoomDto } from "./roomDto";

export interface RoomsListResponse {
  success: boolean;
  data: RoomDto[];
  meta: PaginationMeta;
}